pyproject.toml
[project]
name = "crono"
version = "0.1.0"
description = "A lightweight Python library for scheduling tasks using cron with decorator-based API"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Tom Clesius", email = "tomclesius@gmail.com" }
]
requires-python = ">=3.8"
dependencies = [
    "python-crontab>=3.2.0",
]
keywords = ["cron", "scheduling", "tasks", "async", "decorators"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crono"]

[dependency-groups]
testing = [
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
]

[project.urls]
Homepage = "https://github.com/tclesius/crono"
Repository = "https://github.com/tclesius/crono.git"
Issues = "https://github.com/tclesius/crono/issues"